refactor: splitting growing file
